+++ Framework/Interactive Tutorials/Editor/Editor Windows/TutorialModalWindow.cs
@@ -1,192 +1,155 @@
 using System;
-using System.Collections.Generic;
+using System.Linq;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.UIElements;
+
+using static Unity.InteractiveTutorials.RichTextParser;
 
 namespace Unity.InteractiveTutorials
 {
     // A modal/utility window. Utilizes masking for the modality.
     class TutorialModalWindow : EditorWindow
     {
-        const int kWidth = 700;
-        const int kHeight = 500;
-        const int kLeftColumnWidth = 300;
+        const string k_UIAssetPath = "Packages/com.unity.learn.iet-framework/Framework/UIElementsViews";
+        const int k_Width = 700;
+        const int k_Height = 500;
 
-        [SerializeField]
-        TutorialStyles m_Styles = null;
+        static TutorialStyles Styles => TutorialProjectSettings.instance.TutorialStyle;
+
+        public TutorialWelcomePage WelcomePage
+        {
+            get => m_WelcomePage;
+            set
+            {
+                if (m_WelcomePage)
+                    m_WelcomePage.Modified -= OnWelcomePageModified;
+
+                m_WelcomePage = value;
+                if (m_WelcomePage)
+                    m_WelcomePage.Modified += OnWelcomePageModified;
+            }
+        }
         [SerializeField]
         TutorialWelcomePage m_WelcomePage;
-        List<TutorialParagraphView> m_Paragraphs = new List<TutorialParagraphView>();
-        Action onClose;
+
+        Action m_OnClose;
 
         public static bool Visible { get; private set; }
 
         // Remember to set prior to calling TryToShow().
         public static bool MaskingEnabled { get; set; } = false;
 
-        public static void TryToShow(string windowTitle, TutorialWelcomePage welcomePage, Action onClose)
+        bool IsInitialized => rootVisualElement.childCount > 0;
+
+        static bool s_IsBeingModified;
+        //string m_PreviousWindowTitle;
+
+        public static void TryToShow(TutorialWelcomePage welcomePage, Action onClose)
         {
-            if (Visible)
-                return;
+            var window = EditorWindowUtils.FindOpenInstance<TutorialModalWindow>();
+            if (window)
+                window.Close();
+            window = CreateInstance<TutorialModalWindow>();
+            window.titleContent.text = welcomePage.WindowTitle;
+            //window.m_PreviousWindowTitle = welcomePage.WindowTitle;
+            window.minSize = window.maxSize = new Vector2(k_Width, k_Height);
+            window.m_OnClose = onClose;
+            window.WelcomePage = welcomePage;
 
-            var window = GetWindow<TutorialModalWindow>(utility: true, windowTitle);
-            window.onClose = onClose;
-            var pos = window.position;
-            window.minSize = window.maxSize = new Vector2(kWidth, kHeight);
-            EditorWindowUtils.CenterOnMainWindow(window);
-
-            window.m_WelcomePage = welcomePage;
-            var styles = window.m_Styles;
-
-            foreach (var paragraph in window.m_WelcomePage.paragraphs)
-            {
-                window.m_Paragraphs.Add(
-                    new TutorialParagraphView(paragraph, window, styles.orderedListDelimiter, styles.unorderedListBullet, -1)
-                );
-            }
-
-            window.Show();
+            window.ShowUtility();
+            // NOTE: positioning must be done after Show() in order to work.
+            if (!s_IsBeingModified)
+                EditorWindowUtils.CenterOnMainWindow(window);
 
             if (MaskingEnabled)
                 window.Mask();
         }
 
+        void Initialize()
+        {
+            var windowAsset = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>($"{k_UIAssetPath}/WelcomeDialog.uxml");
+            var styleSheet = AssetDatabase.LoadAssetAtPath<StyleSheet>($"{k_UIAssetPath}/Main.uss");
+            var mainContainer = windowAsset.CloneTree().Q("MainContainer");
+            mainContainer.styleSheets.Add(styleSheet);
+            // TODO OnGuiToolbar is functional, uncomment if/when we reintroduce masking for welcome dialog.
+            //var imguiToolBar = new IMGUIContainer(OnGuiToolbar);
+            //rootVisualElement.Add(imguiToolBar);
+            rootVisualElement.Add(mainContainer);
+        }
+
         void OnEnable()
         {
+            if (!IsInitialized)
+                Initialize();
+        }
+
+        void OnBecameVisible()
+        {
             Visible = true;
-            var root = this.rootVisualElement;
-            //IMGUIContainer iMGUIContainer = new IMGUIContainer(OnGUIOld);
-            //root.Add(iMGUIContainer);
-            VisualTreeAsset welcomeDialogAsset = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>("Packages/com.unity.learn.iet-framework/Framework/UIElementsViews/WelcomeDialog.uxml");
-            StyleSheet rootstyle = AssetDatabase.LoadAssetAtPath<StyleSheet>("Packages/com.unity.learn.iet-framework/Framework/UIElementsViews/WelcomeDialog.uss");
-            VisualElement rootContainer = welcomeDialogAsset.CloneTree().Q("MainContainer");
-            rootContainer.styleSheets.Add(rootstyle);
-            Button GetStartedButton = rootContainer.Q<Button>("GetStartedButton");
-            m_WelcomePage = UserStartupCode.FindWelcomePage();
-
-            GetStartedButton.clicked += Close;
-            GetStartedButton.text = m_WelcomePage.m_StartButtonLabel;
-
-            ((Label)rootContainer.Q("TextContainer").ElementAt(1)).text = m_WelcomePage.title;
-            int currentElement = 2;
-            foreach (TutorialParagraph paragraph in m_WelcomePage.paragraphs)
-            {
-                ((Label)rootContainer.Q("TextContainer").ElementAt(currentElement)).text = paragraph.InstructionTitle;
-                currentElement++;
-            }
-
-
-            if (m_WelcomePage != null && m_WelcomePage.icon != null)
-                rootContainer.Q("LeftColumn").style.backgroundImage = Background.FromTexture2D((Texture2D)m_WelcomePage.icon);
-
-            root.Add(rootContainer);
+            UpdateContent();
             //Mask();
         }
+
+        // For the teardown callbacks the order of execution is OnBecameVisible, OnDisable, OnDestroy.
+        // NOTE OnBecameInvisible appears to be called never if window was shown as utility window.
+        // void OnBecameInvisible() {}
 
         void OnDestroy()
         {
             Visible = false;
-            onClose?.Invoke();
+            s_IsBeingModified = false;
+            m_OnClose?.Invoke();
             Unmask();
         }
 
-        void Update()
+        void UpdateContent()
         {
-            // Force repaint so that changes to WelcomePage can be previed immediately.
-            if (ProjectMode.IsAuthoringMode())
-                Repaint();
-        }
-
-        void OnGUIOld()
-        {
-            if (m_Styles == null)
+            if (!WelcomePage)
             {
-                TutorialStyles.DisplayErrorMessage("TutorialModalWindow.cs");
+                Debug.LogError("null WelcomePage.");
                 return;
             }
 
-            if (m_WelcomePage == null)
-            {
-                return;
-            }
+            var header = rootVisualElement.Q("HeaderMedia");
+            header.style.backgroundImage = Background.FromTexture2D(WelcomePage.Image);
+            rootVisualElement.Q("HeaderContainer").style.display = WelcomePage.Image != null ? DisplayStyle.Flex : DisplayStyle.None;
+            titleContent.text = WelcomePage.WindowTitle;
+            rootVisualElement.Q<Label>("Heading").text = WelcomePage.Title;
+            RichTextToVisualElements(WelcomePage.Description, rootVisualElement.Q("Description"));
 
-            if (Event.current.isKey && Event.current.keyCode == KeyCode.Escape)
-            {
-                Close();
-                return;
-            }
-
-            if (ProjectMode.IsAuthoringMode())
-            {
-                DrawToolbar();
-            }
-
-            GUISkin oldSkin = GUI.skin;
-            GUI.skin = m_Styles.skin;
-
-            using (new EditorGUILayout.HorizontalScope(GUILayout.Width(kWidth), GUILayout.Height(kHeight)))
-            {
-                // left column, "image column"
-                using (new EditorGUILayout.VerticalScope(GUILayout.Width(kLeftColumnWidth), GUILayout.Height(kHeight)))
-                {
-                    GUILayout.Label(GUIContent.none);
-                }
-                if (m_WelcomePage.icon != null)
-                {
-                    GUI.DrawTexture(GUILayoutUtility.GetLastRect(), m_WelcomePage.icon, ScaleMode.StretchToFill);
-                }
-
-                // right column
-                using (new EditorGUILayout.HorizontalScope(AllTutorialStyles.background))
-                {
-                    GUILayout.Space(8f);
-
-                    using (new EditorGUILayout.VerticalScope(AllTutorialStyles.background, GUILayout.Height(kHeight)))
+            var buttonContainer = rootVisualElement.Q("ButtonContainer");
+            buttonContainer.Clear();
+            WelcomePage.Buttons
+                .Where(buttonData => buttonData.Text.Value.IsNotNullOrEmpty())
+                .Select(buttonData =>
+                    new Button(() => buttonData.OnClick?.Invoke())
                     {
-                        const bool pageCompleted = false;
-                        var previousTaskState = true;
-                        foreach (var paragraph in m_Paragraphs)
-                        {
-                            if (paragraph.paragraph.type == ParagraphType.Instruction)
-                                GUILayout.Space(2f);
-
-                            paragraph.Draw(ref previousTaskState, pageCompleted);
-                        }
-
-                        using (new EditorGUILayout.HorizontalScope())
-                        {
-                            if (GUILayout.Button(m_WelcomePage.startButtonLabel, AllTutorialStyles.welcomeDialogButton))
-                            {
-                                Close();
-                            }
-                            GUILayout.FlexibleSpace();
-                        }
-                    }
-                }
-            }
-
-            GUI.skin = oldSkin;
+                        text = buttonData.Text,
+                        tooltip = buttonData.Tooltip
+                    })
+                .ToList()
+                .ForEach(button => buttonContainer.Add(button));
         }
 
-        private void DrawToolbar()
+        void OnGuiToolbar()
         {
             EditorGUILayout.BeginHorizontal(EditorStyles.toolbar, GUILayout.ExpandWidth(true));
 
             GUILayout.FlexibleSpace();
 
             EditorGUI.BeginChangeCheck();
-
             MaskingEnabled = GUILayout.Toggle(
                 MaskingEnabled, "Masking", EditorStyles.toolbarButton,
                 GUILayout.MaxWidth(TutorialWindow.s_AuthoringModeToolbarButtonWidth)
             );
             if (EditorGUI.EndChangeCheck())
             {
-                if (MaskingEnabled) Mask();
-                else Unmask();
+                if (MaskingEnabled)
+                    Mask();
+                else
+                    Unmask();
                 GUIUtility.ExitGUI();
-                return;
             }
 
             EditorGUILayout.EndHorizontal();
@@ -194,11 +157,11 @@ namespace Unity.InteractiveTutorials
 
         void Mask()
         {
-            var styles = m_Styles;
-            var maskingColor = styles?.maskingColor ?? Color.magenta * new Color(1f, 1f, 1f, 0.8f);
-            var highlightColor = styles?.highlightColor ?? Color.cyan * new Color(1f, 1f, 1f, 0.8f);
-            var blockedInteractionColor = styles?.blockedInteractionColor ?? new Color(1, 1, 1, 0.5f);
-            var highlightThickness = styles?.highlightThickness ?? 3f;
+            var styles = Styles;
+            var maskingColor = styles?.MaskingColor ?? Color.magenta * new Color(1f, 1f, 1f, 0.8f);
+            var highlightColor = styles?.HighlightColor ?? Color.cyan * new Color(1f, 1f, 1f, 0.8f);
+            var blockedInteractionColor = styles?.BlockedInteractionColor ?? new Color(1, 1, 1, 0.5f);
+            var highlightThickness = styles?.HighlightThickness ?? 3f;
 
             var unmaskedViews = new UnmaskedView.MaskData();
             unmaskedViews.AddParentFullyUnmasked(this);
@@ -220,6 +183,24 @@ namespace Unity.InteractiveTutorials
         {
             MaskingManager.Unmask();
             MaskingEnabled = false;
+        }
+
+        void OnWelcomePageModified()
+        {
+            s_IsBeingModified = true;
+
+            // TODO try to find a way to author window title for utility window at real-time.
+            //if (WelcomePage.WindowTitle != m_PreviousWindowTitle)
+            //{
+            //    // The way this window is shown currently requires us to recreate the window
+            //    // in order to see a change in the title.
+            //    Close();
+            //    EditorApplication.delayCall += () => TryToShow(WelcomePage, m_OnClose);
+            //}
+            //else
+            {
+                UpdateContent();
+            }
         }
     }
 }