+++ Framework/Interactive Tutorials/Editor/Editor Windows/TutorialWindow.cs
@@ -9,7 +9,7 @@ using UnityEngine.UIElements;
 using Unity.EditorCoroutines.Editor;
 
 using UnityObject = UnityEngine.Object;
-
+using static Unity.InteractiveTutorials.RichTextParser;
 
 namespace Unity.InteractiveTutorials
 {
@@ -504,142 +504,6 @@ namespace Unity.InteractiveTutorials
             EditorCoroutineUtility.StartCoroutineOwnerless(DelayedOnEnable());
         }
 
-        // Transforms html tags to word element labels with different styles to enable rich text
-        static void RichTextToVisualElements(string htmlText, VisualElement tutorialContainerElement)
-        {
-            htmlText = Localization.Tr(htmlText);
-            VisualElementStyleSheetSet style = tutorialContainerElement.styleSheets;
-            tutorialContainerElement.Clear();
-            bool boldOn = false; // <b> sets this on </b> sets off
-            bool italicOn = false; // <i> </i>
-            bool linkOn = false;
-            string linkURL = "";
-            bool firstLine = true;
-            bool lastLineHadText = false;
-            // start streaming text per word to elements while retaining current style for each word block
-            string[] lines = htmlText.Split(new[] { "\r\n", "\r", "\n" }, StringSplitOptions.None);
-
-            foreach (string line in lines)
-            {
-                string[] words = line.Split(new[] { " " }, StringSplitOptions.None);
-
-                if (!firstLine && lastLineHadText)
-                {
-                    AddParagraphToElement(tutorialContainerElement);
-                    lastLineHadText = false;
-                }
-                foreach (string word in words)
-                {
-                    if (word == "" || word == " " || word == "   ") continue;
-                    lastLineHadText = true;
-                    string strippedWord = word;
-                    bool removeBold = false;
-                    bool removeItalic = false;
-                    bool addParagraph = false;
-                    bool removeLink = false;
-
-                    if (strippedWord.Contains("<b>"))
-                    {
-                        strippedWord = strippedWord.Replace("<b>", "");
-                        boldOn = true;
-                    }
-                    if (strippedWord.Contains("<i>"))
-                    {
-                        strippedWord = strippedWord.Replace("<i>", "");
-                        italicOn = true;
-                    }
-                    if (strippedWord.Contains("<a"))
-                    {
-                        strippedWord = strippedWord.Replace("<a", "");
-                        linkOn = true;
-                    }
-                    if (linkOn && strippedWord.Contains("href="))
-                    {
-                        strippedWord = strippedWord.Replace("href=", "");
-                        int linkFrom = strippedWord.IndexOf("\"", StringComparison.Ordinal) + 1;
-                        int linkTo = strippedWord.LastIndexOf("\"", StringComparison.Ordinal);
-                        linkURL = strippedWord.Substring(linkFrom, linkTo - linkFrom);
-                        strippedWord = strippedWord.Substring(linkTo + 2, (strippedWord.Length - 2) - linkTo);
-                        strippedWord.Replace("\">", "");
-                    }
-                    if (strippedWord.Contains("</a>"))
-                    {
-                        strippedWord = strippedWord.Replace("</a>", "");
-                        // TODO </a>text -> also text part is still blue. Parse - for now we can take care when authoring.
-                        removeLink = true;
-                    }
-                    if (strippedWord.Contains("<br/>"))
-                    {
-                        strippedWord = strippedWord.Replace("<br/>", "");
-                        addParagraph = true;
-                    }
-                    if (strippedWord.Contains("</b>"))
-                    {
-                        strippedWord = strippedWord.Replace("</b>", "");
-                        removeBold = true;
-                    }
-                    if (strippedWord.Contains("</i>"))
-                    {
-                        strippedWord = strippedWord.Replace("</i>", "");
-                        removeItalic = true;
-                    }
-                    if (boldOn)
-                    {
-                        Label wordLabel = new Label(strippedWord);
-                        wordLabel.style.color = Color.black;
-                        wordLabel.style.unityFontStyleAndWeight = new StyleEnum<FontStyle>(FontStyle.Bold);
-                        tutorialContainerElement.Add(wordLabel);
-                    }
-                    else if (italicOn)
-                    {
-                        Label wordLabel = new Label(strippedWord);
-                        wordLabel.style.color = Color.black;
-                        wordLabel.style.unityFontStyleAndWeight = new StyleEnum<FontStyle>(FontStyle.Italic);
-                        tutorialContainerElement.Add(wordLabel);
-                    }
-                    else if (addParagraph)
-                    {
-                        AddParagraphToElement(tutorialContainerElement);
-                    }
-                    else if (linkOn && !string.IsNullOrEmpty(linkURL))
-                    {
-                        Label newLabel = new Label(strippedWord);
-
-                        newLabel.style.color = Color.blue;
-                        newLabel.style.borderBottomWidth = 1f;
-                        newLabel.style.borderBottomColor = Color.blue;
-                        newLabel.tooltip = linkURL;
-                        newLabel.RegisterCallback<MouseUpEvent>((evt) => Application.OpenURL(linkURL));
-                        tutorialContainerElement.Add(newLabel);
-                    }
-                    else
-                    {
-                        Label newlabel = new Label(strippedWord);
-                        newlabel.style.color = Color.black;
-                        tutorialContainerElement.Add(newlabel);
-                    }
-                    if (removeBold) boldOn = false;
-                    if (removeItalic) italicOn = false;
-                    if (removeLink)
-                    {
-                        linkOn = false;
-                        linkURL = "";
-                    }
-                }
-                firstLine = false;
-            }
-        }
-
-        static void AddParagraphToElement(VisualElement elementTo)
-        {
-            Label wordLabel = new Label(" ");
-            wordLabel.style.color = Color.black;
-            wordLabel.style.flexDirection = FlexDirection.Row;
-            wordLabel.style.flexGrow = 1f;
-            wordLabel.style.width = 3000f;
-            elementTo.Add(wordLabel);
-        }
-
         void ExitClicked(MouseUpEvent mouseup)
         {
             SkipTutorial();
@@ -876,7 +740,7 @@ namespace Unity.InteractiveTutorials
                     {
                         ++instructionIndex;
                     }
-                    m_AllParagraphs.Add(new TutorialParagraphView(paragraph, instance, styles.orderedListDelimiter, styles.unorderedListBullet, instructionIndex));
+                    m_AllParagraphs.Add(new TutorialParagraphView(paragraph, instance, styles.OrderedListDelimiter, styles.UnorderedListBullet, instructionIndex));
                 }
             }
         }
@@ -1176,7 +1040,11 @@ namespace Unity.InteractiveTutorials
 
             try
             {
-                if (maskingSettings != null && maskingSettings.enabled)
+                if (maskingSettings == null || !maskingSettings.enabled)
+                {
+                    MaskingManager.Unmask();
+                }
+                else
                 {
                     bool foundAncestorProperty;
                     var unmaskedViews = UnmaskedView.GetViewsAndRects(maskingSettings.unmaskedViews, out foundAncestorProperty);
@@ -1224,11 +1092,11 @@ namespace Unity.InteractiveTutorials
 
                     MaskingManager.Mask(
                         unmaskedViews,
-                        styles == null ? Color.magenta * new Color(1f, 1f, 1f, 0.8f) : styles.maskingColor,
+                        styles == null ? Color.magenta * new Color(1f, 1f, 1f, 0.8f) : styles.MaskingColor,
                         highlightedViews,
-                        styles == null ? Color.cyan * new Color(1f, 1f, 1f, 0.8f) : styles.highlightColor,
-                        styles == null ? new Color(1, 1, 1, 0.5f) : styles.blockedInteractionColor,
-                        styles == null ? 3f : styles.highlightThickness
+                        styles == null ? Color.cyan * new Color(1f, 1f, 1f, 0.8f) : styles.HighlightColor,
+                        styles == null ? new Color(1, 1, 1, 0.5f) : styles.BlockedInteractionColor,
+                        styles == null ? 3f : styles.HighlightThickness
                     );
                 }
             }