+++ Framework/Interactive Tutorials/Editor/Localization/LocalizableString.cs
@@ -24,6 +24,8 @@ namespace Unity.InteractiveTutorials
         string m_Untranslated;
 
         public string Translated { get; set; }
+        // TODO returns Translated when localization issues fixed
+        public string Value => /*Translated.AsNullIfEmpty() ??*/ Untranslated;
 
         public LocalizableString() : this(string.Empty) {}
         public LocalizableString(string untranslated) { Untranslated = untranslated; }
@@ -33,9 +35,29 @@ namespace Unity.InteractiveTutorials
         /// Implicit conversion to string returns the translated string, if exists, untranslated otherwise.
         /// </summary>
         /// <param name="str"></param>
-        public static implicit operator string(LocalizableString str) => str.Translated.AsNullIfEmpty() ?? str.Untranslated;
+        public static implicit operator string(LocalizableString str) => str.Value;
 
         public const string PropertyPath = "m_Untranslated";
         public const string OldPropertyPath = "<Untranslated>k__BackingField";
+    }
+
+    // Same as TextAreaAttribute but used for LocalizableStrings.
+    [AttributeUsage(AttributeTargets.Field, Inherited = true, AllowMultiple = false)]
+    public class LocalizableTextAreaAttribute : PropertyAttribute
+    {
+        public readonly int MinLines;
+        public readonly int MaxLines;
+
+        public LocalizableTextAreaAttribute()
+        {
+            MinLines = 3;
+            MaxLines = 3;
+        }
+
+        public LocalizableTextAreaAttribute(int minLines, int maxLines)
+        {
+            MinLines = minLines;
+            MaxLines = maxLines;
+        }
     }
 }