+++ Framework/Interactive Tutorials/Editor/Models/TutorialProjectSettings.cs
@@ -37,21 +37,33 @@ namespace Unity.InteractiveTutorials
             s_Instance = null;
         }
 
+        [Header("Initial Scene and Camera Settings")]
+        [SerializeField]
+        [Tooltip("If set, this page is shown in the welcome dialog when the project is started for the first time.")]
+        TutorialWelcomePage m_WelcomePage = default;
+        public TutorialWelcomePage WelcomePage => m_WelcomePage;
+
         [SerializeField]
         [Tooltip("Initial scene that is loaded when the project is started for the first time.")]
         SceneAsset m_InitialScene = null;
         public SceneAsset initialScene => m_InitialScene;
 
         [SerializeField]
+        SceneViewCameraSettings m_InitialCameraSettings = new SceneViewCameraSettings();
+        public SceneViewCameraSettings InitialCameraSettings => m_InitialCameraSettings;
+
+        [Header("Start-Up Settings")]
+        [SerializeField]
         bool m_RestoreDefaultAssetsOnTutorialReload = true;
-        public bool restoreDefaultAssetsOnTutorialReload { get { return m_RestoreDefaultAssetsOnTutorialReload; } }
+        public bool restoreDefaultAssetsOnTutorialReload => m_RestoreDefaultAssetsOnTutorialReload;
 
         [SerializeField]
-        [Tooltip("If enabled, disregard startup tutorial and start the first tutorial found in the project")]
-        bool m_UseLegacyStartupBehavior = true;
+        [Tooltip("If enabled, disregard startup tutorial and start the first tutorial found in the project.")]
+        bool m_UseLegacyStartupBehavior = default;
 
         [SerializeField]
-        Tutorial m_StartupTutorial = null;
+        [Tooltip("If set, this is the tutorial that can be started from the welcome dialog.")]
+        Tutorial m_StartupTutorial = default;
 
         public Tutorial startupTutorial
         {